MES FINANCIAL DOCS SENT.vbs
CAAD note script that writes what docs were sent to the CP and NCP for a establishment.  Also will create a worklist in CAWT if the CP or the NCP's financial stmt checkboxes are checked and use the return date as the worklist due date.

Problems I cannot resolve:
1) @ line 130 and 140, when only the CP Fin Stmt checkbox is checked, it is creating a worklist for both the CP and the NCP Fin Stmts due.  If the CP Fin Stmt is checked, it should create a worklist just for the CP Fin Stmt due.  When the NCP Fin Stmt is checked, it should create a worklist for the NCP Fin Stmt due.  If both checked then want it to create both the worklists.
2) When neither the CP Fin Stmt or the NCP Fin Stmt checkbox is checked, the worklist part is screwed up.
3) @ line 117, I cannot get the script to write 'Docs Sent to NCP' in the CAAD note after whatever has been checked on the CP side, I want that title to be written before any of the checkboxes on the NCP side are written.

I think that is it!  Let me know where I went wrong, do I need an ELSEIF somewhere???