Create ACTION. UPDATED ESTB NPA DORD DOCS.vbs
This script prints the NPA DORD docs for the CP and NCP of the Financial Statements, Important Statement of Rights and Waivers of Personal Service to the parties. It also adds a CAAD note of what was sent out and a CAWT note to watch for the return of the docs.