#635 Affidavit of Service by Mail Docs Enhancement
Enhanced the script so that it pulls in if there is an attorney on file for the parties and the date of withdrawal if one. Also pulls in if there are Safety Concerns, Added checkbox to click is service is done by the direction of the County Attorney. Lastly it will update automatically the county in which service is being done and the name of the worker on the case.